+++ src/main/java/eu/eutampieri/catacombs/model/Entity.java
@@ -84,7 +84,7 @@ public abstract class Entity extends GameObject implements LivingCharacter {
      * @param delta time between updates
      */
     @Override
-    public void update(final int delta) {
+    public void update(final float delta) {
         move();
         updateSpriteLocation();
     }
