+++ src/main/java/eu/eutampieri/catacombs/ui/StartTransition.java
@@ -19,7 +19,7 @@ public class StartTransition extends State{
     }
 
     @Override
-    public void update(final int delta) {
+    public void update(final float delta) {
         this.waitTimer += delta;
 
         if (this.waitTimer >= WAIT_TIME) {
